Investigate random read errors

+++ src/Sensor.h
@@ -2,7 +2,7 @@
 class Sensor {
 public:
   Sensor(int triggerPin, int echoPin);
-  float getDistance();
+  int getDistance();
 
 private:
   int triggerPin;
